Revert "Merge branch '1.20.2-server-renderer' into 1.20.2"
This reverts commit 56466c01b42616280d18f968f84687b115cb8e2c, reversing
changes made to 453e36af5a0192e985f142b96381efb6b609f918.

+++ src/main/java/xiamomc/morph/backends/server/renderer/network/datawatcher/values/CatValues.java
@@ -9,7 +9,7 @@ public class CatValues extends TameableAnimalValues
     public final SingleValue<CatVariant> CAT_VARIANT = getSingle("cat_variant", BuiltInRegistries.CAT_VARIANT.getOrThrow(CatVariant.TABBY));
     public final SingleValue<Boolean> IS_LYING = getSingle("cat_is_lying", false);
     public final SingleValue<Boolean> RELAXED = getSingle("cat_relaxed", false);
-    public final SingleValue<Byte> COLLAR_COLOR = getSingle("cat_collar_color", (byte)14);
+    public final SingleValue<Integer> COLLAR_COLOR = getSingle("cat_collar_color", 14);
 
     public CatValues()
     {
